requirements.txt
django >=3.2,<=3.3
python-decouple
djangorestframework
markdown    
django-filter